clamp cobbler at the version I started with in early 2008
especially demo 07 is affected if you don't do this.  Taking that
as an example, you can always find some other, more recent, date
to use in the --since there, but it just seems simpler this way.

Keeping up with cobbler is irrelevant for this demo; it's just a
convenient example...